requirements-gpu.txt
ipykernel
numpy
tqdm
opencv-python
googledrivedownloader
requests
tensorflow-gpu==1.13.1
albumentations==0.4.5
Flask
Flask-JWT-Extended
Flask-RESTful
